+++ recall-store/src/main/java/com/aitusoftware/recall/map/CharSequenceMap.java
@@ -71,7 +71,7 @@ public final class CharSequenceMap implements SequenceMap<CharSequence>
         totalEntryCount = BitUtil.findNextPositivePowerOfTwo(initialSize);
         entryMask = totalEntryCount - 1;
         entrySizeInBytes = (maxKeyLength * Character.BYTES) + (4 * Integer.BYTES);
-        final long bufferSize = entrySizeInBytes * totalEntryCount;
+        final long bufferSize = entrySizeInBytes * (long)totalEntryCount;
         if (bufferSize > ((long)Integer.MAX_VALUE) || totalEntryCount < 0)
         {
             throw new IllegalArgumentException("Requested buffer size too large");
